Rename snapshot export package directory
