--- cloud/disk_manager/internal/pkg/dataplane/snapshot/export/tests/ya.make
@@ -1,3 +1,0 @@
-GO_TEST_FOR(cloud/disk_manager/internal/pkg/dataplane/snapshot/export)
-
-END()
