+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/export.go
