+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/export_test.go
