+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/partition_helpers_test.go
