+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/partition_integrity_test.go
