+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/partition_map_test.go
