+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/tests/ya.make
@@ -1,0 +1,3 @@
+GO_TEST_FOR(cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter)
+
+END()
