+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/exporter/ya.make
