+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/ya.make
@@ -10,6 +10,6 @@ END()
 
 RECURSE(
     config
-    export
+    exporter
     storage
 )
